+++ client/src/pages/LocationPage/LeftColumnDetails.tsx
@@ -7,9 +7,8 @@ import CameraAltOutlinedIcon from '@mui/icons-material/CameraAltOutlined';
 import ScreenShareOutlinedIcon from '@mui/icons-material/ScreenShareOutlined';
 import BookmarkBorderOutlinedIcon from '@mui/icons-material/BookmarkBorderOutlined';
 import type { LocationType } from '../../types/LocationType';
-import CommentForm from './CommentSection/Form/CommentForm';
+import CommentForm from './CommentSection/CommentForm';
 import { useState } from 'react';
-
 
 function LeftColumnDetais({ locationInfo }: { locationInfo: LocationType }) {
   const [showCommentForm, setShowCommentForm] = useState(false);
@@ -27,7 +26,12 @@ function LeftColumnDetais({ locationInfo }: { locationInfo: LocationType }) {
           >
             Write a comment
           </Button>
-          <CommentForm locationName={locationInfo.name} locationType={locationInfo.type} isVisible={showCommentForm} onClose={() => setShowCommentForm(false)}/> 
+          <CommentForm
+            locationName={locationInfo.name}
+            locationType={locationInfo.type}
+            isVisible={showCommentForm}
+            onClose={() => setShowCommentForm(false)}
+          />
         </div>
         <div className="location-page-left-button">
           <Button
